Additional refactor for better readability, better typing for syncToUrl:False
